ratectrl_qmode_test: fix use of uniform_int_distribution
behavior is only defined for short->unsigned long long, not uint8_t:

https://en.cppreference.com/w/cpp/numeric/random/uniform_int_distribution
IntType	-	The result type generated by the generator. The effect
is undefined if this is not one of short, int, long, long long, unsigned
short, unsigned int, unsigned long, or unsigned long long.

this fixes a build error under Visual Studio:
random(2401): error C2338: invalid template argument for
uniform_int_distribution: N4659 29.6.1.1 [rand.req.genl]/1e requires one
of short, int, long, long long, unsigned short, unsigned int, unsigned
long, or unsigned long long

+++ test/ratectrl_qmode_test.cc
@@ -1139,8 +1139,7 @@ TEST_F(RateControlQModeTest, TestKMeans) {
   std::default_random_engine generator;
   for (int &centroid : centroids_ref) {
     // This is to make sure each cluster is far enough from others.
-    std::uniform_int_distribution<uint8_t> distribution(centroid - 8,
-                                                        centroid + 8);
+    std::uniform_int_distribution<int> distribution(centroid - 8, centroid + 8);
     for (int i = 0; i < num_sample_per_cluster; ++i) {
       const int random_sample = distribution(generator);
       random_input.push_back(random_sample);
